requirements.txt
psutil==5.9.8
gputil==1.4.0
requests==2.31.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.15.0
matplotlib==3.8.2
pandas==2.2.0